use output dir, use hyperfine_iteration (WIP)
Introduce a mandatory `--output-dir` option when running benchmarks.
This will, in the future, be used to store benchmark configuration,
results and run artifacts.

Currently it will store configuration and the final run artifact, as the
changes needed in hyperfine to expose the run iteration are not
available yet. (see: https://github.com/sharkdp/hyperfine/pull/807 )